Type checks continued.

+++ tests/files/basics.mys.hpp
@@ -28,7 +28,7 @@ void func_4(void);
     constexpr auto __name__ = [] (auto &&...args) { \
         return mys::basics::func_4(std::forward<decltype(args)>(args)...); \
     };
-List<i32> func_5(void);
+List<i64> func_5(void);
 #define MYS_BASICS_func_5_IMPORT_AS(__name__) \
     constexpr auto __name__ = [] (auto &&...args) { \
         return mys::basics::func_5(std::forward<decltype(args)>(args)...); \
